# Run our custom eslint rules using ts-node