--- a/.vscode/settings.json
+++ b/.vscode/settings.json
@@ -41,11 +41,6 @@
 			}
 		}
 	],
-	"eslint.options": {
-		"rulePaths": [
-			"./build/lib/eslint"
-		]
-	},
 	"typescript.tsdk": "node_modules/typescript/lib",
 	"npm.exclude": "**/extensions/**",
 	"npm.packageManager": "yarn",