--- a/build/lib/eslint-plugin-vscode/code-import-patterns.ts
+++ b/build/lib/eslint-plugin-vscode/code-import-patterns.ts
@@ -6,7 +6,7 @@
 import * as eslint from 'eslint';
 import { TSESTree } from '@typescript-eslint/experimental-utils';
 import * as path from 'path';
-import * as minimatch from 'minimatch';
+import minimatch from 'minimatch';
 import { createImportRuleListener } from './utils';
 
 const REPO_ROOT = path.normalize(path.join(__dirname, '../../../'));