--- a/build/lib/eslint-plugin-vscode/index.js
+++ b/build/lib/eslint-plugin-vscode/index.js
@@ -1,0 +1,12 @@
+const glob = require('glob');
+const path = require('path');
+
+require('ts-node').register({ experimentalResolver: true });
+
+// Re-export all .ts files as rules
+const rules = {};
+glob.sync(`${__dirname}/*.ts`).forEach((file) => {
+	rules[path.basename(file, '.ts')] = require(file);
+});
+
+module.exports = { rules };