--- a/build/lib/eslint-plugin-vscode/index.js
+++ b/build/lib/eslint-plugin-vscode/index.js
@@ -1,7 +1,7 @@
 const glob = require('glob');
 const path = require('path');
 
-require('ts-node').register({ experimentalResolver: true });
+require('ts-node').register({ experimentalResolver: true, transpileOnly: true });
 
 // Re-export all .ts files as rules
 const rules = {};