--- a/build/lib/eslint-plugin-vscode/package.json
+++ b/build/lib/eslint-plugin-vscode/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@vscode/eslint-plugin",
+  "private": true,
+  "version": "0.0.0",
+  "main": "index.js"
+}