--- a/build/package.json
+++ b/build/package.json
@@ -62,7 +62,6 @@
     "source-map": "0.6.1",
     "through2": "^4.0.2",
     "tmp": "^0.2.1",
-    "ts-node": "^10.9.1",
     "vsce": "^1.100.0",
     "vscode-universal-bundler": "^0.0.2"
   },