--- a/build/tsconfig.build.json
+++ b/build/tsconfig.build.json
@@ -7,5 +7,8 @@
 	},
 	"include": [
 		"**/*.ts"
+	],
+	"exclude": [
+		"lib/eslint-plugin-vscode/**/*"
 	]
 }