--- a/test/monaco/esm-check/index.js
+++ b/test/monaco/esm-check/index.js
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-// eslint-disable-next-line code-no-standalone-editor
+// eslint-disable-next-line @vscode/code-no-standalone-editor
 import * as monaco from './out/vs/editor/editor.main.js';
 
 monaco.editor.create(document.getElementById('container'), {