@jrieken 2022-08-16
LGTM but I believe there is specific build performance reason to have these files as JS in the repo. Adding @joaomoreno and @alexdima to comment on that