@mjbvz 2022-08-17
If performance is a concern, we could also try enabling ts-node's `transpileOnly` option or use `swc` to switch to using that compiler. This would skip the expensive type checking process