@joaomoreno 2022-08-24
Yeah, it was a performance choice. So, seeing runtime numbers here would help.