@mjbvz 2022-08-28
@joaomoreno Some quick performance tests using `time yarn eslint`

- Main: `yarn run eslint  101.58s user 5.47s system 119% cpu 1:29.77 total`

- With TS Node: `yarn run eslint  108.45s user 5.33s system 124% cpu 1:31.41 total`

- With TS Node (transpile only): `yarn run eslint  107.12s user 5.03s system 123% cpu 1:30.46 total`

- With TS Node (swc): `yarn run eslint  101.08s user 4.93s system 124% cpu 1:25.26 total`

There may be some small performance gains from switching from eslint's `rulesDir` like we do on main, to using a plugin like this branch does