@joaomoreno 2022-08-29
Yeah `transpileOnly` seems like the best balance here. But we'll still be type checking these scripts in some CI job, right?