@mjbvz 2022-08-29
@joaomoreno Type checking is done by the `npmCheckJs` script under `build`. However that is currently not run as part of CI. If we think running this makes sense, I'll take a look at enabling it in a separate PR